+++ src/app/portlet/terra-portlet.component.ts
@@ -14,8 +14,6 @@ import {
     trigger
 } from "@angular/animations";
 import { isNullOrUndefined } from 'util';
-import { TerraButtonComponent } from '../button/terra-button.component';
-import { forEach } from '@angular/router/src/utils/collection';
 import { TerraButtonInterface } from '../button/data/terra-button.interface';
 
 @Component({