+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Feature
 * **terra-portlet** add new input `inputButtonList` to add a button group into the portlet header
+* **terra-button** add properties `isSmall` and `isLarge` to button interface
 
 <a name="2.1.39-zlk"></a>
 # 2.1.39-zlk (05.01.2017)